% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{example_taxa}
\alias{example_taxa}
\title{Example taxonomy assignments}
\usage{
example_taxa()
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} in the format used by \code{\link[=taxtable]{taxtable()}} giving a
simple set of taxonomy assignments suitable for use in \code{\link[=phylotax]{phylotax()}}.
Tip labels are the same as those used in \code{\link[=example_tree]{example_tree()}}.
}
\description{
Example taxonomy assignments
}
\examples{
example_taxa()
}
