% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotax.R
\name{example_tree}
\alias{example_tree}
\title{Simple phylogenetic tree for use in examples}
\usage{
example_tree()
}
\value{
a \code{\link[ape:read.tree]{phylo}} object giving a very simple tree;
tip labels are the same as those in \code{\link[=example_taxa]{example_taxa()}}.
}
\description{
Simple phylogenetic tree for use in examples
}
\examples{
example_tree()
}
