% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotax.R
\name{extract_taxon}
\alias{extract_taxon}
\title{Extract a taxon by name from a \code{phylotax} object}
\usage{
extract_taxon(phylotax, taxon, true_members = NULL)
}
\arguments{
\item{phylotax}{(\link[=phylotax]{phylotax} object) The object to extract from.}

\item{taxon}{(\code{character} string) The name of the taxon to extract.}

\item{true_members}{(\code{character}) (optional) Labels of sequences known to
belong in the taxon.}
}
\value{
A \link[=phylotax]{phylotax} object containing only the specified taxon.
}
\description{
For a \link[=phylotax]{phylotax} object with a tree, the taxon is extracted based
on node annotations.  If more than one node is annotated as belonging to the
given taxon (i.e., it seems to be para/polyphyletic in the tree) then the
"correct" clade can be chosen by providing "true members"; the result will
only include the clade(s) which contain the "true members".
}
\details{
If no tree is present (meaning the \code{phylotax} object was generated using
\code{\link[=lca_consensus]{lca_consensus()}}), and also for sequences which are not present in the tree,
then the chosen taxon is simply extracted using the assigned taxonomy.
}
