% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotax.R
\name{lca_consensus}
\alias{lca_consensus}
\title{Last common ancestor consensus}
\usage{
lca_consensus(
  taxa,
  ranks = NULL,
  method = if (utils::hasName(taxa, "method")) "LCA" else NULL
)
}
\arguments{
\item{taxa}{(\code{data.frame}) Taxon assignments for the taxa on the tree,
as returned by \code{taxtable()}.  Should have columns "\code{label}",
"\code{rank}", and "\code{taxon}", but may have other columns as well; see the \code{method}
argument.  Multiple assignments for each label are allowed, and
can be generated by using \code{rbind()} or \code{dplyr::bind_rows()}
on results from multiple calls to \code{taxonomy()} and
\code{taxtable()}.}

\item{ranks}{(\code{character}) names of ranks
used in the taxon assignments, from most to least inclusive. This can be
omitted if the "\code{rank}" column of "\code{taxa}" is an \code{integer} or ordered
\code{factor}, in which case the \emph{smallest} value is the \emph{most inclusive} rank, or
alternatively if the "\code{rank}" column of "\code{taxa}" is a \code{character} or \code{factor}
with values/levels from the set rootrank, domain, kingdom, phylum, class, order, family, genus, species.}

\item{method}{(a single \code{character} string, or a named \code{character} vector)
how to identify different methods. See details.}
}
\value{

}
\description{
This is an alternative to \code{\link[=phylotax]{phylotax()}} which does not require a phylogenetic
tree. Instead, multiple assignments are resolved by strict consensus at each
rank in turn.  For eaxample, if for a particular sequence all available
assignments at the kingdom, phylum, and class agree, but there is
disagreement at the order level, then the consensus assignment takes the
kingdom, phylum, and class assignments to be correct, but does not include
any assignment at the order level or below.
}
\section{Distinguishing different primary methods}{
Primary methods can be distinguished in three ways:
\enumerate{
\item Not at all.  For this option, \code{taxa} should not have a column named
"\code{method}", and the the \code{method} argument to \code{lca_consensus()} should be
\code{NULL}.
\item A single column named "\code{method}" in \code{taxa}. If \code{method=NULL} but \code{taxa}
has a "\code{method}" column, then this column is assumed to uniquely
identify the methods. Assignments made by \code{lca_consensus()} will have
"\code{consensus}" in the method column. This value can be changed by setting
the \code{method} argument to an unnamed \code{character} string, e.g.,
"\code{method = 'cons'}".
\item Custom columns. If the \code{method} argument is a named character vector,
then the names are taken to be columns in \code{taxa} (which must exist) and
the values are taken to be the values for each column which should be
used for consensus annotations, e.g.,
"\code{method = c(algorithm = "consensus", region = "ITS2")}". Each unique
combination of values in these columns is treated as a distinct method.
}
}

