% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{phylotax}
\alias{phylotax}
\title{Assign taxon labels to nodes in a tree when there is a consensus of IDs on descendent tips.}
\usage{
phylotax(
  tree = ape::read.tree(text = paste0("(", paste(unique(taxa$label), collapse = ","),
    ");")),
  taxa
)
}
\arguments{
\item{tree}{(\code{\link[ape]{phylo}} object) A tree including the taxa to be
classified.  The tip labels should match the "label" column in \code{taxa}.}

\item{taxa}{(\code{data.frame}) Taxon assignments for the taxa on the tree,
as returned by \code{\link{taxtable}}.  Should have columns "label",
"rank", and "taxon", but may have other columns as well.  Multiple
assignments for each lable are allowed, and
can be generated by using \code{\link{rbind}} or \code{\link[dplyr]{bind_rows}}
on results from multiple calls to \code{\link{taxonomy}} and
\code{\link{taxtable}}.}
}
\value{
a list with two elements, "tip_taxa" and "node_taxa".  "tip_taxa" is
    a \code{\link[tibble]{tibble}} with the same format as \code{taxa}, in
    which assignments which are inconsistent with the phylogeny have been
    removed, and new assignments deduced or confirmed from the phylogeny.
    These are identified by the value "phylotax" in the "method" column,
    which is created if it does not already exist.  "node_taxa" has columns
    "node", "rank" and "taxon", giving taxonomic assignments for the nodes of
    the tree.
}
\description{
Assign taxon labels to nodes in a tree when there is a consensus of IDs on descendent tips.
}
