% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phylotax_to_phyloseq}
\alias{phylotax_to_phyloseq}
\title{Convert assigned taxonomy and tree to a phyloseq object}
\usage{
phylotax_to_phyloseq(phylotax, otu_table, ..., use_tree = TRUE)
}
\arguments{
\item{phylotax}{(\code{\link[=phylotax]{phylotax}} object)}

\item{otu_table}{(\code{\link[phyloseq:otu_table-class]{phyloseq::otu_table}} object)}

\item{...}{Additional arguments to pass to \code{\link[phyloseq:phyloseq]{phyloseq::phyloseq()}}. These
should generally be the result of \code{\link[phyloseq:sample_data]{phyloseq::sample_data()}}, but could also
include a \code{\link[ape:read.tree]{phylo}} object if \code{phylotax} does not already
have one or if \code{use_tree=FALSE}.}

\item{use_tree}{(\code{logical}) If \code{FALSE}, do not use the tree from \code{phylotax}
even if one exists.}
}
\value{
A \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}
object.
}
\description{
The output will have \code{\link[phyloseq:tax_table]{tax_table}} defined based on
the \emph{assigned} taxa in \code{phylotax}, and \code{\link[phyloseq:phy_tree]{phy_tree}}
defined based on the tree in \code{phylotax}, if one is present.
}
\details{
The resulting object will contain all taxon labels found in \emph{both}
\code{otu_table} and the tree associated with \code{phylotax}, if any. Taxon labels
with no assignments in \code{phylotax} will have \code{NA} for all ranks in the
output's taxonomy table.
}
