% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{relabel_tree}
\alias{relabel_tree}
\title{Replace tree tip labels}
\usage{
relabel_tree(tree, old, new, quote = FALSE, drop = character(0))
}
\arguments{
\item{tree}{(\link[ape:read.tree]{phylo} object) phylogenetic tree with tip labels}

\item{old}{(\code{character}) Tip labels from \code{tree} which should be changed.
This does not have to be all the labels in the tree.}

\item{new}{(\code{character}) Replacement tip labels.  Should have the same length
as \code{old}.}

\item{quote}{(\code{logical}) If TRUE, double quotes are added around the values
in \code{new} before they are used for replacement.  This is helpful if they (may)
contain characters which will must be quoted in Newick format.}

\item{drop}{(\code{character}) (optional) Tip labels which should be dropped
before replacement.}
}
\value{
The tree with modified tip labels.
}
\description{
Replace tree tip labels
}
