% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotax-package.R
\docType{data}
\name{sebacinales_oldnames}
\alias{sebacinales_oldnames}
\title{Original genus-level names of the Sebacinales dataset}
\format{
An object of class \code{character} of length 54.
}
\source{
https://rdp.cme.msu.edu/
}
\usage{
sebacinales_oldnames
}
\description{
Original genus-level names of the Sebacinales dataset
}
\keyword{datasets}
