% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\alias{taxonomy_dada2}
\alias{taxonomy_sintax}
\alias{taxonomy_idtaxa}
\title{Assign taxonomy to nucleotide sequences}
\usage{
taxonomy(
  seq.table,
  reference,
  method,
  min_confidence = 50,
  multithread = FALSE,
  ...
)

taxonomy_dada2(
  seq,
  reference,
  multithread = FALSE,
  tryRC = FALSE,
  outputBootstraps = TRUE,
  verbose = TRUE,
  ...
)

taxonomy_sintax(
  seq,
  reference,
  min_confidence = NULL,
  multithread = FALSE,
  exec = "vsearch",
  ...
)

taxonomy_idtaxa(
  seq,
  reference,
  multithread = FALSE,
  strand = "top",
  min_confidence = 40,
  ...
)
}
\arguments{
\item{seq.table}{(\code{character} vector or something that can be coerced to
one, or a matrix with sequences as the column names) Sequences to
assign taxonomy}

\item{reference}{(\code{character} string giving a path to a file or the
result from \code{\link[DECIPHER]{LearnTaxa}}) An appropriately
formatted reference database}

\item{method}{(\code{character} string) taxonomy assignment method.
Currently accepted values are "dada2", "sintax", and "idtaxa".}

\item{min_confidence}{(\code{integer} between 0 and 100) The minimum
confidence to report results.}

\item{multithread}{(\code{integer} scalar) the number of processors to use
for assignment.}

\item{...}{additional arguments for methods}

\item{tryRC}{(\code{logical} scalar) passed on to \code{\link[dada2]{assignTaxonomy}}}

\item{outputBootstraps}{(\code{logical} scalar) passed on to \code{\link[dada2]{assignTaxonomy}}}

\item{verbose}{(\code{logical} scalar) passed on to \code{\link[dada2]{assignTaxonomy}}}

\item{exec}{(\code{character} string) name of the executable to use for
SINTAX search.  The default is "vsearch", but "usearch" should also
work.  In either case, the executable should be installed and on the
system path.}

\item{strand}{(\code{character} string) passed on to \code{\link[DECIPHER]{IdTaxa}}}
}
\value{
raw results of the taxonomy assignment, of various types depending on
       \code{method}.  (\code{method = "dada2"}) returns a \code{list} with
       elements "tax" and "boot", as \code{\link[dada2]{assignTaxonomy}}.
       (\code{method = "sintax"}) returns a \code{data.frame} with columns
       "label", "hit", "strand", and, if \code{min_confidence} is not given,
       "c12n" (short for "classification"). (\code{method = "idtaxa"}) gives
       an S4 object of classes "Taxa" and "Train".
       Any of these can be passed to \code{\link{taxtable}} to get a
       uniform format.
}
\description{
This method dispatches to other R packages or external programs.
}
