% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxtable}
\alias{taxtable}
\title{Convert results from different taxonomic assignment algorithms to a uniform
format}
\usage{
taxtable(tax, ...)
}
\arguments{
\item{tax}{Results from \code{\link{taxonomy}}}

\item{...}{passed to methods}

\item{min_confidence}{(\code{integer}) The minimum confidence to include in
results. May be higher than the value given in \code{\link{taxonomy}},
but will have no effect if it is lower.}
}
\value{
a \code{\link[tibble]{tibble}} with columns: \describe{
  \item{\code{label}}{sequence identifier}
  \item{\code{rank}}{the rank of the assignment}
  \item{\code{taxon}}{the taxon which was assigned}
  \item{\code{confidence}}{the confidence of the assignment}
} Each query sequence will typically occupy several rows of the output, one
for each rank which was assigned.
}
\description{
Convert results from different taxonomic assignment algorithms to a uniform
format
}
