% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotax.R
\name{phylotax}
\alias{phylotax}
\title{Assign taxon labels to nodes in a tree when there is a consensus of IDs on
descendant tips.}
\usage{
phylotax(
  tree = NULL,
  taxa,
  ranks = NULL,
  method = if (utils::hasName(taxa, "method")) "PHYLOTAX" else NULL
)
}
\arguments{
\item{tree}{(\code{ape::phylo()} object) A tree including the taxa to be
classified.  The tip labels should match the "label" column in \code{taxa}. If no
tree is given, a star tree will be used, which results in each tip being
assigned to the strict consensus of its primary assignments, if any.}

\item{taxa}{(\code{data.frame}) Taxon assignments for the taxa on the tree,
as returned by \code{taxtable()}.  Should have columns "label",
"rank", and "taxon", but may have other columns as well; see the \code{method}
argument.  Multiple assignments for each label are allowed, and
can be generated by using \code{rbind()} or \code{dplyr::bind_rows()}
on results from multiple calls to \code{taxonomy()} and
\code{taxtable()}.}

\item{ranks}{(\code{character}) names of ranks
used in the taxon assignments, from most to least inclusive. This can be
omitted if the "\code{rank}" column of "\code{taxa}" is an \code{integer} or ordered \code{factor},
in which case the \emph{smallest} value is the \emph{most inclusive} rank, or
alternatively if the "\code{rank}" column of "\code{taxa}" is a \code{character} or \code{factor}
with values/levels from the set rootrank, domain, kingdom, phylum, class, order, family, genus, species.}

\item{method}{(a single \code{character} string, or a named \code{character} vector)
how to identify different methods. See details.}
}
\value{
an S3 object with class "\code{phylotax}", with five elements:
\itemize{
\item "\code{tip_taxa} a \code{tibble::tibble()} with the same format as \code{taxa}, containing
taxonomy assignments made by PHYLOTAX to tips.
\item "\code{node_taxa}" a \code{tibble::tibble()} with columns "\code{node}", "\code{label}",
"\code{rank}" and "\code{taxon}" giving taxonomy assignments made by PHYLOTAX to
internal nodes.
\item "\code{rejected}" a \code{tibble::tibble()} with the same format as \code{taxa} giving
primary assignments which have been rejected by PHYLOTAX.
\item "\code{retained}" a \code{tibble::tibble()} with the same format as \code{taxa} giving
primary assignments which have not been rehected by PHYLOTAX. These may
contain inconsistencies that PHYLOTAX was unable to resolve.
\item "\code{missing}" a \code{tibble::tibble()} with the same format as \code{taxa}, giving the
primary assignments which have not been assessed by PHULOTAX because they
have labels which are not present on the tree.
}
}
\description{
Assign taxon labels to nodes in a tree when there is a consensus of IDs on
descendant tips.
}
\section{Distinguishing different primary methods}{
Primary methods can be distinguished in three ways:
\enumerate{
\item Not at all.  For this option, \code{taxa} should not have a column named
"\code{method}", and the the \code{method} argument to \code{phylotax()} should be
\code{NULL}.
\item A single column named "\code{method}" in \code{taxa}. If \code{method=NULL} but \code{taxa}
has a "\code{method}" column, then this column is assumed to uniquely
identify the methods. Assignments made by \code{phylotax()} will have
"\code{PHYLOTAX}" in the method column. This value can be changed by setting
the \code{method} argument to an unnamed \code{character} string, e.g.,
"\code{method = 'consensus'}".
\item Custom columns. If the \code{method} argument is a named character vector,
then the names are taken to be columns in \code{taxa} (which must exist) and
the values are taken to be the values for each column which should be
used for PHYLOTAX annotations, e.g.,
"\code{method = c(algorithm = "PHYLOTAX", region = "ITS2")}". PHYLOTAX will
treat each unique combination of values in these columns as a distinct
method.
}
}

