% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{phylotax}
\alias{phylotax}
\title{Assign taxon labels to nodes in a tree when there is a consensus of IDs on descendent tips.}
\usage{
phylotax(
  tree = ape::read.tree(text = paste0("(", paste(unique(taxa$label), collapse = ","),
    ");")),
  taxa,
  ranks = NULL
)
}
\arguments{
\item{tree}{(\code{ape::phylo()} object) A tree including the taxa to be
classified.  The tip labels should match the "label" column in \code{taxa}.}

\item{taxa}{(\code{data.frame}) Taxon assignments for the taxa on the tree,
as returned by \code{taxtable()}.  Should have columns "label",
"rank", and "taxon", but may have other columns as well.  Multiple
assignments for each label are allowed, and
can be generated by using \code{rbind()} or \code{dplyr::bind_rows()}
on results from multiple calls to \code{taxonomy()} and
\code{taxtable()}.}

\item{ranks}{(\code{character}) names of ranks
used in the taxon assignments, from most to least inclusive. This can be
omitted if the "\code{rank}" column of "\code{taxa}" is an \code{integer} or ordered \code{factor},
in which case the \emph{smallest} value is the \emph{most inclusive} rank, or
alternatively if the "\code{rank}" column of "\code{taxa}" is a \code{character} or \code{factor}
with values/levels from the set rootrank, domain, kingdom, phylum, class, order, family, genus, species.}
}
\value{
a list with two elements, "tip_taxa" and "node_taxa".  "tip_taxa" is
a \code{tibble::tibble()} with the same format as \code{taxa}, in
which assignments which are inconsistent with the phylogeny have been
removed, and new assignments deduced or confirmed from the phylogeny.
These are identified by the value "phylotax" in the "method" column,
which is created if it does not already exist.  "node_taxa" has columns
"node", "rank" and "taxon", giving taxonomic assignments for the nodes of
the tree.
}
\description{
Assign taxon labels to nodes in a tree when there is a consensus of IDs on descendent tips.
}
